Moved AuditMiddleware to Core, Removed unused files
