+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Core/Models/Claims/OrgClaim.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Altinn.AccessMgmt.PersistenceEF.Models.Claims
+namespace Altinn.AccessMgmt.Core.Models.Claims
 {
     /// <summary>
     /// Organization claim matching structure from maskinporten
