--- src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.PersistenceEF/Configurations/Base/IBasicDbConfiguration.cs
@@ -1,3 +1,0 @@
-﻿namespace Altinn.AccessMgmt.PersistenceEF.Configurations.Base;
-
-public interface IBasicDbConfiguration { }
